+++ Landform/GeometryCommand.cs
@@ -20,7 +20,7 @@ namespace OPS.Landform
         public virtual int TextureResolution { get; set; }
 
         [Option(HelpText = "Max texture atlas stretch (0 = no stretch, 1 = unlimited)", Default = 1)]
-        public virtual float MaxTextureStretch { get; set; }
+        public virtual double MaxTextureStretch { get; set; }
     }
 
     public class GeometryCommand : WedgeCommand
@@ -30,7 +30,7 @@ namespace OPS.Landform
         protected Mesh mesh; //finest LOD
         protected string meshFrame;
         protected int sceneTextureResolution;
-        protected float maxTextureStretch;
+        protected double maxTextureStretch;
 
         public GeometryCommand(GeometryCommandOptions gcopts) : base(gcopts)
         {
